+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+
+#### 0.80
+
+- Add check for well known names used by malware sandboxes.
+- Improve ProcessDebugObject anti-debug check thanks to @Mattiwatti
+
+
 #### 0.79
 - Add anti-disassembly trick: Jump with constant condition
 - Add anti-disassembly trick: Jump instruction with same target